requirements.txt
alembic==1.4.3
click==7.1.2
dependency-injector==4.5.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
nose==1.3.7
psycopg2==2.8.4
PyHamcrest==2.0.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
